refactor(server): split chat handler and deduplicate LLM/VLM paths
Break the 947-line chat.go into focused files (messages, response,
stream, token) and deduplicate the LLM/VLM paths behind a generic
runChat with a prepareFn. Token handling collapses to a single
tokenClass classifier shared by the block sink and stream render
adapters, replacing four parallel plain/reasoning functions.

No behavior change: the default inline output stays byte-identical.
